requirements.txt
kivy
requests
python-dateutil
